refactor: InsertStatement를 local class로 변경
